Streamlines VPK tool execution in build
Removes `DotNetToolRunner` and `VpkToolResolver`, directly packages the `vpk` executable in `Velopack.Build.csproj`, updates `PackTask` and `PublishTask` to invoke the embedded `vpk`, and extracts shared process argument handling to `ProcessExtensions.Shared.cs`.
