+++ src/vpk/Velopack.Build/Velopack.Build.csproj
@@ -6,28 +6,62 @@
     <LangVersion>13</LangVersion>
     <SignAssembly>false</SignAssembly>
     <NoWarn>$(NoWarn);NU5128</NoWarn>
+    <GenerateNuspecDependsOn>$(GenerateNuspecDependsOn);IncludeVpkInPackage</GenerateNuspecDependsOn>
+    <TargetsForTfmSpecificContentInPackage>$(TargetsForTfmSpecificContentInPackage);IncludeTfmDependenciesInPackage</TargetsForTfmSpecificContentInPackage>
   </PropertyGroup>
 
   <PropertyGroup>
     <VelopackPackageId>Velopack.Build</VelopackPackageId>
     <VelopackPackageVendorLibs>true</VelopackPackageVendorLibs>
-    <VelopackPackageSelfContained>true</VelopackPackageSelfContained>
-    <DevelopmentDependency>true</DevelopmentDependency>
+    <IncludeBuildOutput>false</IncludeBuildOutput>
+    <!-- TODO: we might be able to remove the entire SelfContained.targets file now -->
+    <VelopackPackageSelfContained>false</VelopackPackageSelfContained>
+    <!-- Removed DevelopmentDependency to allow package dependencies -->
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\lib-csharp\Util\ProcessExtensions.cs" Link="ProcessExtensions.cs" />
+    <Compile Include="..\..\lib-csharp\Util\ProcessExtensions.Shared.cs" Link="ProcessExtensions.Shared.cs" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Build.Utilities.Core" Version="18.6.3" />
     <PackageReference Include="Microsoft.IO.Redist" Version="6.1.3" Condition="'$(TargetFramework)' == 'net472'" />
     <PackageReference Include="Riok.Mapperly" Version="4.3.1" />
+    <PackageReference Include="Microsoft.Build.Utilities.Core" Version="17.14.28" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.IO.Redist" Version="6.1.3" Condition="'$(TargetFramework)' == 'net472'" PrivateAssets="All" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
+    <ProjectReference Include="..\Velopack.Vpk\Velopack.Vpk.csproj">
+      <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
+      <PrivateAssets>all</PrivateAssets>
+    </ProjectReference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\lib-csharp\Velopack.csproj">
+      <PrivateAssets>none</PrivateAssets>
+    </ProjectReference>
   </ItemGroup>
 
   <ItemGroup>
     <None Include="Velopack.Build.targets" Pack="true" PackagePath="build/Velopack.Build.targets" />
     <None Include="Velopack.Build.props" Pack="true" PackagePath="build/Velopack.Build.props" />
   </ItemGroup>
+
+  <!-- Include Velopack.Vpk output in the tools directory -->
+  <Target Name="IncludeVpkInPackage" BeforeTargets="Build">
+    <ItemGroup>
+      <None Include="$(MSBuildThisFileDirectory)..\..\..\build\$(Configuration)\net8.0\publish\**\*.*" Pack="true" PackagePath="tools\net8.0\" />
+    </ItemGroup>
+  </Target>
+
+  <Target Name="IncludeTfmDependenciesInPackage">
+    <ItemGroup>
+      <TfmSpecificPackageFile Include="$(TargetDir)**/*.*">
+        <PackagePath>build\$(TargetFramework)</PackagePath>
+      </TfmSpecificPackageFile>
+    </ItemGroup>
+  </Target>
 
 </Project>
