+++ src/vpk/Velopack.Build/Velopack.Build.props
@@ -1,1 +1,4 @@
-<Project><PropertyGroup><VelopackToolMode Condition="'$(VelopackToolMode)' == ''">Auto</VelopackToolMode><VelopackToolVersion Condition="'$(VelopackToolVersion)' == ''"></VelopackToolVersion><VelopackToolPrerelease Condition="'$(VelopackToolPrerelease)' == ''">false</VelopackToolPrerelease><VelopackToolSource Condition="'$(VelopackToolSource)' == ''"></VelopackToolSource><VelopackSkipToolInstall Condition="'$(VelopackSkipToolInstall)' == ''">false</VelopackSkipToolInstall></PropertyGroup></Project>
+<Project>
+  <PropertyGroup>
+  </PropertyGroup>
+</Project>
