+++ src/vpk/Velopack.Build/Velopack.Build.targets
@@ -4,7 +4,7 @@
     <VelopackPushOnPublish Condition="$(VelopackPushOnPublish) == ''">false</VelopackPushOnPublish>
     <!-- https://learn.microsoft.com/visualstudio/msbuild/tutorial-custom-task-code-generation?view=vs-2022&WT.mc_id=DT-MVP-5003472#changes-required-to-multitarget -->
     <VelopackStronglyTyped_TFM Condition=" '$(MSBuildRuntimeType)' != 'Core' ">net472</VelopackStronglyTyped_TFM>
-    <VelopackStronglyTyped_TFM Condition=" '$(MSBuildRuntimeType)' == 'Core' ">net6.0</VelopackStronglyTyped_TFM>
+    <VelopackStronglyTyped_TFM Condition=" '$(MSBuildRuntimeType)' == 'Core' ">net8.0</VelopackStronglyTyped_TFM>
     <VelopackBuildDll>$(MSBuildThisFileDirectory)$(VelopackStronglyTyped_TFM)\Velopack.Build.dll</VelopackBuildDll>
   </PropertyGroup>
 
@@ -93,11 +93,6 @@
       Categories="$(VelopackCategories)"
       Shortcuts="$(VelopackShortcuts)"
       Compression="$(VelopackAppImageCompression)"
-      VelopackToolMode="$(VelopackToolMode)"
-      VelopackToolVersion="$(VelopackToolVersion)"
-      VelopackToolPrerelease="$(VelopackToolPrerelease)"
-      VelopackToolSource="$(VelopackToolSource)"
-      VelopackSkipToolInstall="$(VelopackSkipToolInstall)"
     />
   </Target>
 
@@ -110,11 +105,6 @@
       Channel="$(VelopackChannel)"
       Timeout="$(VelopackHttpTimeout)"
       WaitForLive="$(VelopackWaitForLive)"
-      VelopackToolMode="$(VelopackToolMode)"
-      VelopackToolVersion="$(VelopackToolVersion)"
-      VelopackToolPrerelease="$(VelopackToolPrerelease)"
-      VelopackToolSource="$(VelopackToolSource)"
-      VelopackSkipToolInstall="$(VelopackSkipToolInstall)"
     />
   </Target>
 </Project>
