--- test/Velopack.Build.Tests/GlobalUsings.cs
@@ -1,8 +1,0 @@
-﻿global using System;
-global using System.Collections.Generic;
-global using System.IO;
-global using System.Linq;
-global using System.Threading;
-global using System.Threading.Tasks;
-global using Xunit;
-global using Xunit.Abstractions;
